Added Stemming/Lemmatization
Now, you can enter in an unlemmatized translation dict and it with
lemmatize/stem in (stem the french, lemmatize the English, just due to
the tools available).  We may want to just use this on the outside of
this code and then directly use the lemmatized dictionary in the code,
rather than having the code do this every time we run it.