Fixed Translator
DirectTranslate had the following issues that are fixed in this version.
1. Accidentally called "stem" on the english lemmatized instead of
"lemmatize"
2. Didn't stem the search word
3. Didn't convert the search word to lower case
4. Encoding of file wasn't utf

Also added a note to the README about the necessary NLTK packages and
corpora to download.

Also changed the test code to go on real french sentences.

NOTE: There still seem to be some encoding issues.  I suspect they are
with the dictionary and lookup.